feat(conditionBuilder): spellchecks , comments , fixes
