+++ packages/ibm-products-styles/src/components/ConditionBuilder/_condition-builder.scss
@@ -32,7 +32,7 @@ $block-class: #{c4p-settings.$pkg-prefix}--condition-builder;
   display: flex;
 }
 
-.#{$block-class}__condtion-wrapper {
+.#{$block-class}__condition-wrapper {
   display: flex;
   flex-flow: row wrap;
 }
