+++ packages/ibm-products/src/components/ConditionBuilder/ConditionBuilderContent/ConditionBuilderContent.js
@@ -19,13 +19,13 @@ const blockClass = `${pkg.prefix}--condition-builder`;
 
 function ConditionBuilderContent({ startConditionLabel, conditionBuilderRef }) {
   const { rootState, setRootState } = useContext(ConditionBuilderContext);
-  const [isConditionbuilderActive, setIsConditionbuilderActive] =
+  const [isConditionBuilderActive, setIsConditionBuilderActive] =
     useState(true);
   const conditionBuilderContentRef = useRef();
 
   useEffect(() => {
     if (rootState?.groups?.length) {
-      setIsConditionbuilderActive(false);
+      setIsConditionBuilderActive(false);
       if (
         rootState.groups[0].conditions.length == 1 &&
         rootState.groups[0].conditions[0].property == undefined
@@ -38,12 +38,12 @@ function ConditionBuilderContent({ startConditionLabel, conditionBuilderRef }) {
         }, 0);
       }
     } else {
-      setIsConditionbuilderActive(true);
+      setIsConditionBuilderActive(true);
     }
   }, [rootState]);
 
   useEffect(() => {
-    if (isConditionbuilderActive == false) {
+    if (isConditionBuilderActive == false) {
       if (conditionBuilderRef.current) {
         const initial =
           conditionBuilderRef.current.querySelector('[role="row"]');
@@ -52,10 +52,10 @@ function ConditionBuilderContent({ startConditionLabel, conditionBuilderRef }) {
         }
       }
     }
-  }, [isConditionbuilderActive, conditionBuilderRef]);
-  const onStartConditionbuilder = () => {
+  }, [isConditionBuilderActive, conditionBuilderRef]);
+  const onStartConditionBuilder = () => {
     //when add condition button is clicked.
-    setIsConditionbuilderActive(false);
+    setIsConditionBuilderActive(false);
     setRootState(emptyState); //here we can set an empty skeleton object for an empty condition builder,
     //or we can even pre-populate some existing builder and continue editing
   };
@@ -77,14 +77,14 @@ function ConditionBuilderContent({ startConditionLabel, conditionBuilderRef }) {
       className={`${blockClass}__content-container`}
       tabIndex={-1}
     >
-      {isConditionbuilderActive && (
+      {isConditionBuilderActive && (
         <Button
           className={`${blockClass}__condition-builder`}
           renderIcon={(props) => <Add size={16} {...props} />}
           iconDescription={startConditionLabel}
           kind="ghost"
           size="sm"
-          onClick={onStartConditionbuilder}
+          onClick={onStartConditionBuilder}
         >
           {startConditionLabel}
         </Button>
