+++ packages/ibm-products/src/components/ConditionBuilder/ConditionBuilderContext/DataConfigs.js
@@ -1,4 +1,5 @@
-import { translationsObject } from './translations';
+import { translationsObject } from './translationObject';
+
 export const statementConfig = [
   {
     label: 'if',
