+++ packages/ibm-products/src/components/ConditionBuilder/ConditionBuilderContext/translationObject.js
@@ -1,0 +1,22 @@
+export const translationsObject = {
+    en: {
+      if: 'if',
+      'excl-if': 'excl.if',
+      and: 'and',
+      or: 'or',
+      is: 'is',
+      greater: 'is greater than',
+      'greater-equal': 'is greater than or equal to',
+      lower: 'is lower than',
+      'lower-equal': 'is lower than or equal to',
+      'starts-with': 'starts with',
+      'ends-with': 'ends with',
+      contains: 'contains',
+      'one-of': 'is one of',
+      before: 'is before',
+      after: 'is after',
+      between: 'is between',
+      'add-condition': 'Add Condition',
+    },
+  };
+
